Refactor: 친구 목록 컴포넌트에 전달하는 props 수정 친한친구 수정 요청 useMutation 적용
